Removed SHIFT button and stop watch (stop watch didn't work because you tried to access the current time of the watch INSIDE a button function, which is only called when a button got pressed. If you would press a button after exactly 10 seconds (to be clear after 10,0 to 10,99999... seconds) it would have worked KIND OF that way you did it. You had to expand the watch.elapsed function by .Seconds which returns you the current SECOND state of the watch. Nevertheless this is a calculator, not a stop-watch, so I removed it... Back to changes: The original shift button now toggles between arc and normal functions (sin/cos/tan) by rearranging the button click events as I told you on the phone.
PLEASE READ: You always use input_length = 0 after an operation. This is not really necessary, because we have the variable "new_operand" which you can set on true in order to have the same result. Please do so. Input length = 0 is only required for the numbers (especially 0 which is not allowed to be entered TWICE at the beginning, because there is no number like 00123 or so).